ospf6d: Support updating the set of MDR routable neighbors immediately.
Add a configuration option to control if the set of MDR routable
neighbors is updated immediately after the shortest-path tree and
routing table are calculated.  When enabled, a second shortest-path
tree and routing table calculation is performed if the set of routable
neighbors has changed.

Currently, the set of routable neighbors is always updated before
sending a hello.  Using this option can allow installing more direct
routes without waiting the hello interval.