RPM: Persist /var/run/quagga
On EL7 and newer, /var/run/quagga is a tmpfs that goes away on reboots.

The current package creates this directory, but on reboot the directory
is gone so the services fail to start. There is a systemd facility
to create files and directories at boot time.

This commit copies the upstream CentOS7 tmpfiles config file,
quagga-tmpfs.conf:
https://git.centos.org/rpms/quagga/blob/c7/f/SOURCES/quagga-tmpfs.conf
and the necessary SPEC lines to install it from the upstream spec:
https://git.centos.org/rpms/quagga/blob/c7/f/SPECS/quagga.spec