vtysh: Limit files in xpimd/ processed by extract.pl
During the build, the script vtysh/extract.pl identifies the supported VTY
commands and their structure, by using the source code itself. To do this,
the script directly calls the C/C++ preprocessor on individual files, and
then it parses the output.

This fails when examining certain files in xpimd/: the preprocessor is not
called with the flags needed to locate all the header files they include.
However, those files do not define VTY commands and can be ignored here.

Resolve this by limiting the files in xpimd/ that are read by this script.